Move remaining DiagramToolbox code in DiagramPage
